+++ test/API/VMTest.cpp
@@ -62,7 +62,7 @@ QBDI_NOINLINE int dummyFunCall(int arg0) {
 
 #if defined(QBDI_ARCH_X86) || defined(QBDI_ARCH_X86_64)
 #define MNEM_COUNT 5u
-#define MNEM_VALIDATION 123u
+#define MNEM_VALIDATION 140u
 #elif defined(QBDI_ARCH_ARM)
 #define MNEM_COUNT 1u
 #define MNEM_VALIDATION 25u
@@ -76,51 +76,54 @@ struct TestInst {
     uint32_t instSize;
     uint8_t numOperands;
     bool isCompare;
+    QBDI::RegisterAccessType flagsAccess;
     QBDI::OperandAnalysis operands[6];
 };
 
 #if defined(QBDI_ARCH_X86) || defined(QBDI_ARCH_X86_64)
 struct TestInst TestInsts[MNEM_COUNT] = {
-    {3, 2, true, {
+    {3, 2, true, QBDI::REGISTER_WRITE, {
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 1, 8, 3, "DH", QBDI::REGISTER_READ},
-           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_NONE, MNEM_IMM_SHORT_VAL, 1, 0, 0, nullptr, QBDI::REGISTER_UNUSED},
+           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_NONE, MNEM_IMM_SHORT_VAL, 1, 0, -1, nullptr, QBDI::REGISTER_UNUSED},
         }
     },
 #if defined(QBDI_ARCH_X86_64)
-    {3, 2, true, {
+    {3, 2, true, QBDI::REGISTER_WRITE, {
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 8, 0, 0, "RAX", QBDI::REGISTER_READ},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 8, 0, 1, "RBX", QBDI::REGISTER_READ},
         }
     },
 #else
-    {3, 2, true, {
+    {3, 2, true, QBDI::REGISTER_WRITE, {
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 2, 0, 0, "AX", QBDI::REGISTER_READ},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 2, 0, 1, "BX", QBDI::REGISTER_READ},
         }
     },
 #endif
-    {5, 2, true, {
-           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_NONE, MNEM_IMM_VAL, 4, 0, 0, nullptr, QBDI::REGISTER_UNUSED},
+    {5, 2, true, QBDI::REGISTER_WRITE, {
+           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_NONE, MNEM_IMM_VAL, 4, 0, -1, nullptr, QBDI::REGISTER_UNUSED},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 4, 0, 0, "EAX", QBDI::REGISTER_READ},
         }
     },
-    {1, 4, false, {
+    {1, 5, false, QBDI::REGISTER_READ | QBDI::REGISTER_WRITE, {
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_ADDR, 0, sizeof(QBDI::rword), 0, 5, QBDI::GPR_NAMES[5], QBDI::REGISTER_READ},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_ADDR, 0, sizeof(QBDI::rword), 0, 4, QBDI::GPR_NAMES[4], QBDI::REGISTER_READ},
+           {QBDI::OPERAND_INVALID, QBDI::OPERANDFLAG_NONE, 0, 0, 0, -1, NULL, QBDI::REGISTER_UNUSED},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 4, 0, 5, "EDI", QBDI::REGISTER_READ_WRITE},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, 4, 0, 4, "ESI", QBDI::REGISTER_READ_WRITE},
         }
     },
 #if defined(QBDI_ARCH_X86_64)
-    {5, 5, true, {
+    {5, 6, true, QBDI::REGISTER_WRITE, {
 #else
-    {4, 5, true, {
+    {4, 6, true, QBDI::REGISTER_WRITE, {
 #endif
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_NONE, 0, sizeof(QBDI::rword), 0, 0, QBDI::GPR_NAMES[0], QBDI::REGISTER_READ},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_ADDR, 0, sizeof(QBDI::rword), 0, 4, QBDI::GPR_NAMES[4], QBDI::REGISTER_READ},
-           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_ADDR, 1, sizeof(QBDI::rword), 0, 0, NULL, QBDI::REGISTER_UNUSED},
+           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_ADDR, 1, sizeof(QBDI::rword), 0, -1, NULL, QBDI::REGISTER_UNUSED},
            {QBDI::OPERAND_GPR, QBDI::OPERANDFLAG_ADDR, 0, sizeof(QBDI::rword), 0, 5, QBDI::GPR_NAMES[5], QBDI::REGISTER_READ},
-           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_ADDR, 3, sizeof(QBDI::rword), 0, 0, NULL, QBDI::REGISTER_UNUSED},
+           {QBDI::OPERAND_IMM, QBDI::OPERANDFLAG_ADDR, 3, sizeof(QBDI::rword), 0, -1, NULL, QBDI::REGISTER_UNUSED},
+           {QBDI::OPERAND_INVALID, QBDI::OPERANDFLAG_NONE, 0, 0, 0, -1, NULL, QBDI::REGISTER_UNUSED},
         }
     },
 };
@@ -128,8 +131,8 @@ struct TestInst TestInsts[MNEM_COUNT] = {
 struct TestInst TestInsts[MNEM_COUNT] = {
     {4, 3, true, {
             {QBDI::OPERAND_GPR, 0, sizeof(QBDI::rword), 0, 3, "R3", QBDI::REGISTER_READ},
-            {QBDI::OPERAND_IMM, MNEM_IMM_SHORT_VAL, sizeof(QBDI::rword), 0, 0, nullptr, QBDI::REGISTER_UNUSED},
-            {QBDI::OPERAND_PRED, 0, sizeof(QBDI::rword), 0, 0, nullptr, QBDI::REGISTER_UNUSED},
+            {QBDI::OPERAND_IMM, MNEM_IMM_SHORT_VAL, sizeof(QBDI::rword), 0, -1, nullptr, QBDI::REGISTER_UNUSED},
+            {QBDI::OPERAND_PRED, 0, sizeof(QBDI::rword), 0, -1, nullptr, QBDI::REGISTER_UNUSED},
         }
     }
 };
@@ -354,6 +357,9 @@ QBDI::VMAction evilMnemCbk(QBDI::VMInstanceRef vm, QBDI::GPRState *gprState, QBD
                     CHECKED_IF(!ana->isReturn)
                         CHECKED_IF(ana->isCompare)
                             info[1]++;
+        }
+        CHECKED_IF(ana->flagsAccess == currentInst.flagsAccess) {
+            info[1]++;
         }
         // validate number of analyzed operands
         CHECKED_IF(ana->numOperands == currentInst.numOperands) {